Figured out a refactor and got tests to pass!
